refactor(hub): extract index maintenance from collection.ts (Phase 5 A15)
