+++ packages/hub/src/collection.ts
@@ -41,6 +41,16 @@ import {
   buildPersistedIndexCallbacks as buildPersistedIndexCallbacksImpl,
   type SearchContext,
 } from './with-lookup/search/collection-facade.js'
+import {
+  rebuildEagerIndexesFromCache as rebuildEagerIndexesFromCacheImpl,
+  rebuildUniqueConstraintsFromCache as rebuildUniqueConstraintsFromCacheImpl,
+  rebuildIndexes as rebuildIndexesImpl,
+  reconcileIndex as reconcileIndexImpl,
+  maintainPersistedIndexesOnPut as maintainPersistedIndexesOnPutImpl,
+  maintainPersistedIndexesOnDelete as maintainPersistedIndexesOnDeleteImpl,
+  purgePersistedIndexes as purgePersistedIndexesImpl,
+  type IndexingContext,
+} from './with-lookup/indexing/collection-facade.js'
 import { ConflictError, ReadOnlyError, TranslatorNotConfiguredError, LocaleNotSpecifiedError } from './errors.js'
 import type { GhostRecord, TierMode, CrossTierAccessEvent } from './types.js'
 import type { UnlockedKeyring } from './with-party/team/keyring.js'
@@ -60,8 +70,8 @@ import { NO_HISTORY, type HistoryStrategy } from './with-commit/history/strategy
 import { Query, ScanBuilder } from './query/index.js'
 import type { QuerySource, JoinContext, JoinableSource } from './query/index.js'
 import type { CollectionIndexes, IndexDef } from './with-lookup/indexing/eager-indexes.js'
-import { encodeIdxId, decodeIdxId } from './with-lookup/indexing/persisted-indexes.js'
-import type { PersistedCollectionIndex, PersistedIndexDef } from './with-lookup/indexing/persisted-indexes.js'
+import { decodeIdxId } from './with-lookup/indexing/persisted-indexes.js'
+import type { PersistedCollectionIndex } from './with-lookup/indexing/persisted-indexes.js'
 import { LazyQuery } from './with-lookup/indexing/lazy-builder.js'
 import type { LazyQuerySource } from './with-lookup/indexing/lazy-builder.js'
 import { NO_INDEXING, type IndexStrategy, type IndexState } from './with-lookup/indexing/strategy.js'
@@ -69,7 +79,7 @@ import type { SearchOptions, SearchResult } from './with-lookup/search/index.js'
 import { MemoryIndexStore, type IndexStore } from './with-lookup/search/index-store.js'
 import { PersistedIndexStore } from './with-lookup/search/persisted-index-store.js'
 import type { RetrieveOptions, RetrieveHit } from './with-lookup/search/retrieve-types.js'
-import { IndexWriteFailureError, DerivationCapExceededError, EmbeddingDimMismatchError } from './errors.js'
+import { DerivationCapExceededError, EmbeddingDimMismatchError } from './errors.js'
 import { embeddingSourceText, VectorSet, type EmbeddingDescriptor } from './with-lookup/embeddings/index.js'
 import { buildUniqueConstraintSet, type UniqueConstraintSet } from './with-lookup/indexing/unique-constraints.js'
 import type { RefDescriptor } from './refs.js'
@@ -4001,13 +4011,7 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
    * 1K–50K records this completes in single-digit milliseconds.
    */
   private rebuildEagerIndexesFromCache(): void {
-    const eager = this.indexes
-    if (!eager || eager.fields().length === 0) return
-    const snapshot: Array<{ id: string; record: T }> = []
-    for (const [id, entry] of this.cache) {
-      snapshot.push({ id, record: entry.record })
-    }
-    eager.build(snapshot)
+    rebuildEagerIndexesFromCacheImpl(this.indexingContext())
   }
 
   /**
@@ -4015,12 +4019,7 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
    * Called after any bulk hydration alongside `rebuildEagerIndexesFromCache`.
    */
   private rebuildUniqueConstraintsFromCache(): void {
-    if (!this.uniqueConstraints) return
-    this.uniqueConstraints.build(
-      (function* (cache: Map<string, { record: T }>) {
-        for (const [id, entry] of cache) yield [id, entry.record] as const
-      })(this.cache),
-    )
+    rebuildUniqueConstraintsFromCacheImpl(this.indexingContext())
   }
 
   /**
@@ -4043,51 +4042,7 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
    * per-field drift repair, use `reconcileIndex(field)` instead.
    */
   async rebuildIndexes(): Promise<void> {
-    if (!this.lazy) {
-      await this.ensureHydrated()
-      this.rebuildEagerIndexesFromCache()
-      return
-    }
-
-    const persisted = this.persistedIndexes
-    if (!persisted) return
-    const fields = persisted.fields()
-    if (fields.length === 0) return
-
-    // 1. Collect canonical ids (skip every reserved-namespace id —
-    //    `_idx/`, `_keyring`, `_history/`, `_ledger_deltas/`, `_meta/`,
-    //    `_ledger`, `_blob_`, etc. User records may not start with `_`
-    //    per the monorepo convention used across the hub).
-    const allIds = await this.adapter.list(this.vault, this.name)
-    const canonicalIds: string[] = []
-    const staleIdxIds: string[] = []
-    for (const id of allIds) {
-      if (decodeIdxId(id)) {
-        staleIdxIds.push(id)
-      } else if (!id.startsWith('_')) {
-        canonicalIds.push(id)
-      }
-    }
-
-    // 2. Drop every existing side-car. Errors here are tolerated — the
-    //    next step overwrites any remnants. If a side-car is for a
-    //    field that is no longer declared, the delete still removes
-    //    the stale row from storage.
-    for (const id of staleIdxIds) {
-      try { await this.adapter.delete(this.vault, this.name, id) } catch { /* ignore */ }
-    }
-    persisted.clear()
-
-    // 3. Walk records and write fresh side-cars for every declared field.
-    for (const recordId of canonicalIds) {
-      const envelope = await this.adapter.get(this.vault, this.name, recordId)
-      if (!envelope) continue
-      const record = await this.codec.decryptRecord(envelope, { skipValidation: true })
-      if (record === null) continue // shredded (tombstone) — no side-car to build
-      await this.maintainPersistedIndexesOnPut(recordId, record, null, envelope._v)
-    }
-
-    this.persistedIndexesLoaded = true
+    return rebuildIndexesImpl(this.indexingContext())
   }
 
   /**
@@ -4112,111 +4067,7 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
     field: string,
     opts: { dryRun?: boolean } = {},
   ): Promise<{ field: string; missing: string[]; stale: string[]; applied: number }> {
-    if (!this.lazy) {
-      throw new Error(
-        `Collection "${this.name}": reconcileIndex is only meaningful in lazy mode ` +
-        `(prefetch: false). Eager mode maintains indexes in memory with no drift.`,
-      )
-    }
-    const persisted = this.persistedIndexes
-    if (!persisted) {
-      throw new Error(
-        `Collection "${this.name}": indexing is disabled on this Noydb instance. ` +
-        `Pass \`withIndexing()\` from "@noy-db/hub/indexing" to \`createNoydb({ indexStrategy })\`.`,
-      )
-    }
-    if (!persisted.has(field)) {
-      throw new Error(
-        `Collection "${this.name}": field "${field}" is not declared in indexes. ` +
-        `Declare it in the collection options before reconciling.`,
-      )
-    }
-
-    const dryRun = opts.dryRun === true
-    const allIds = await this.adapter.list(this.vault, this.name)
-
-    // Map side-car recordId → stored value (if readable). Also capture
-    // "stale" side-cars whose field matches but whose record is gone.
-    const sidecar = new Map<string, unknown>()
-    const sidecarIds = new Map<string, string>() // recordId -> sidecar id
-    for (const id of allIds) {
-      const decoded = decodeIdxId(id)
-      if (!decoded || decoded.field !== field) continue
-      sidecarIds.set(decoded.recordId, id)
-      const env = await this.adapter.get(this.vault, this.name, id)
-      if (!env) continue
-      try {
-        const sidecarJson = await this.codec.decryptJsonString(env)
-        if (sidecarJson === null) {
-          // Tombstone side-car (shredded) — treat as stale so it's rewritten.
-          sidecar.set(decoded.recordId, undefined)
-        } else {
-          const body = JSON.parse(sidecarJson) as { value: unknown }
-          sidecar.set(decoded.recordId, body.value)
-        }
-      } catch {
-        // Unreadable — treat as stale so it gets rewritten.
-        sidecar.set(decoded.recordId, undefined)
-      }
-    }
-
-    // Walk canonical records and compare against side-car state.
-    const missing: string[] = []
-    const stale: string[] = []
-    const fixesPut: Array<{ recordId: string; record: T; version: number }> = []
-    for (const id of allIds) {
-      if (decodeIdxId(id)) continue
-      if (id.startsWith('_')) continue
-      const env = await this.adapter.get(this.vault, this.name, id)
-      if (!env) continue
-      const record = await this.codec.decryptRecord(env, { skipValidation: true })
-      // Shredded (tombstone) canonical record: treat like a vanished record —
-      // leave its `id` in `sidecarIds` so any lingering side-car is marked
-      // stale (and deleted) by the leftover loop below.
-      if (record === null) continue
-      const live = readPersistedValue(record as unknown as Record<string, unknown>, field)
-      const stored = sidecar.get(id)
-      const hasSidecar = sidecarIds.has(id)
-      const indexable = live !== null && live !== undefined
-
-      if (indexable && !hasSidecar) {
-        missing.push(id)
-        fixesPut.push({ recordId: id, record, version: env._v })
-      } else if (indexable && hasSidecar && !valuesMatch(stored, live)) {
-        // Side-car body drifted from live value (e.g. partial write
-        // after an update). Rewrite so lookups agree with reality.
-        missing.push(id)
-        fixesPut.push({ recordId: id, record, version: env._v })
-      } else if (!indexable && hasSidecar) {
-        // Record exists but its value is no longer indexable (null/
-        // undefined). The side-car is stale.
-        stale.push(sidecarIds.get(id)!)
-      }
-      sidecarIds.delete(id)
-    }
-    // Any side-car whose canonical record vanished is stale.
-    for (const [, idxId] of sidecarIds) stale.push(idxId)
-
-    let applied = 0
-    if (!dryRun) {
-      for (const idxId of stale) {
-        try {
-          await this.adapter.delete(this.vault, this.name, idxId)
-          applied++
-        } catch { /* ignore — next reconcile picks it up */ }
-      }
-      for (const fix of fixesPut) {
-        await this.maintainPersistedIndexesOnPut(fix.recordId, fix.record, null, fix.version)
-        applied++
-      }
-      // In-memory mirror is authoritative for query dispatch — make
-      // sure it matches what's on disk now.
-      persisted.clear()
-      this.persistedIndexesLoaded = false
-      await this.ensurePersistedIndexesLoaded()
-    }
-
-    return { field, missing, stale, applied }
+    return reconcileIndexImpl(this.indexingContext(), field, opts)
   }
 
   /**
@@ -4469,57 +4320,13 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
    * `PersistedCollectionIndex.stringifyKey` contract. If the prior value
    * was non-null and the new value is null, the side-car is deleted.
    */
-  private async maintainPersistedIndexesOnPut(
+  private maintainPersistedIndexesOnPut(
     id: string,
     newRecord: T,
     previousRecord: T | null,
     version: number,
   ): Promise<void> {
-    const persisted = this.persistedIndexes
-    if (!persisted) return
-    const defs = persisted.definitions()
-    if (defs.length === 0) return
-
-    const newRec = newRecord as unknown as Record<string, unknown>
-    const prevRec = previousRecord as unknown as Record<string, unknown> | null
-
-    for (const def of defs) {
-      const newValue = extractIndexValue(newRec, def)
-      const previousValue = prevRec ? extractIndexValue(prevRec, def) : null
-
-      // Update the in-memory mirror first — it's the authoritative source
-      // for query dispatch. If the adapter write below fails, the mirror
-      // still reflects intended state; the reconciler compares mirror
-      // against side-cars on next run.
-      persisted.upsert(id, def.key, newValue, previousValue)
-
-      const idxId = encodeIdxId(def.key, id)
-      try {
-        if (newValue === null || newValue === undefined) {
-          // Clear any pre-existing side-car for this (field, record).
-          if (previousValue !== null && previousValue !== undefined) {
-            await this.adapter.delete(this.vault, this.name, idxId)
-          }
-        } else {
-          const body = JSON.stringify({
-            field: def.key,
-            value: serializeIndexValue(newValue),
-            recordId: id,
-            writtenAt: new Date().toISOString(),
-          })
-          const envelope = await this.codec.encryptJsonString(body, version)
-          await this.adapter.put(this.vault, this.name, idxId, envelope)
-        }
-      } catch (cause) {
-        this.emitter.emit('index:write-partial', {
-          vault: this.vault,
-          collection: this.name,
-          id,
-          action: 'put',
-          error: new IndexWriteFailureError({ recordId: id, field: def.key, op: 'put', cause }),
-        })
-      }
-    }
+    return maintainPersistedIndexesOnPutImpl(this.indexingContext(), id, newRecord, previousRecord, version)
   }
 
   /**
@@ -4527,32 +4334,8 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
    * Mirror state updates regardless of adapter outcome; adapter failures
    * surface on `index:write-partial` the same way put does.
    */
-  private async maintainPersistedIndexesOnDelete(id: string, previousRecord: T): Promise<void> {
-    const persisted = this.persistedIndexes
-    if (!persisted) return
-    const defs = persisted.definitions()
-    if (defs.length === 0) return
-
-    const prevRec = previousRecord as unknown as Record<string, unknown>
-    for (const def of defs) {
-      const previousValue = extractIndexValue(prevRec, def)
-      if (previousValue !== null && previousValue !== undefined) {
-        persisted.remove(id, def.key, previousValue)
-      }
-
-      const idxId = encodeIdxId(def.key, id)
-      try {
-        await this.adapter.delete(this.vault, this.name, idxId)
-      } catch (cause) {
-        this.emitter.emit('index:write-partial', {
-          vault: this.vault,
-          collection: this.name,
-          id,
-          action: 'delete',
-          error: new IndexWriteFailureError({ recordId: id, field: def.key, op: 'delete', cause }),
-        })
-      }
-    }
+  private maintainPersistedIndexesOnDelete(id: string, previousRecord: T): Promise<void> {
+    return maintainPersistedIndexesOnDeleteImpl(this.indexingContext(), id, previousRecord)
   }
 
   /**
@@ -4568,20 +4351,8 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
    * stale mirror hit cannot surface the erased record. Returns the count deleted
    * + the `def.key`s whose delete FAILED (residue that still leaks the value).
    */
-  async _purgePersistedIndexes(id: string): Promise<{ purged: number; residue: string[] }> {
-    const persisted = this.persistedIndexes
-    if (!persisted) return { purged: 0, residue: [] }
-    let purged = 0
-    const residue: string[] = []
-    for (const def of persisted.definitions()) {
-      try {
-        await this.adapter.delete(this.vault, this.name, encodeIdxId(def.key, id))
-        purged++
-      } catch {
-        residue.push(def.key)
-      }
-    }
-    return { purged, residue }
+  _purgePersistedIndexes(id: string): Promise<{ purged: number; residue: string[] }> {
+    return purgePersistedIndexesImpl(this.indexingContext(), id)
   }
 
   /** #308 L2 — drop a record's encrypted _vec sidecar on erasure (a vector is text-invertible).
@@ -4886,76 +4657,29 @@ export class Collection<T, S extends keyof T = never, Q extends keyof T & string
       emitCrossTierEvent: (event) => this.emitCrossTierEvent(event),
     }
   }
-}
 
-/**
- * Read a field value from a plain record for persisted-index maintenance.
- * Supports dotted paths so declarations like `indexes: ['billing.clientId']`
- * work the same way `readPath` handles them for the eager-mode builder.
- */
-function readPersistedValue(record: Record<string, unknown>, field: string): unknown {
-  if (!field.includes('.')) return record[field]
-  const segments = field.split('.')
-  let cursor: unknown = record
-  for (const segment of segments) {
-    if (cursor === null || cursor === undefined) return undefined
-    cursor = (cursor as Record<string, unknown>)[segment]
-  }
-  return cursor
-}
-
-/**
- * Canonicalize a typed value for storage inside the side-car body so it
- * round-trips through `JSON.parse` without losing fidelity. Dates are
- * serialised as ISO strings; everything else passes through.
- *
- * The in-memory mirror compares on the stringified bucket key, so the
- * exact storage form is not query-critical — this just protects the
- * reconciler, which compares the stored body against the
- * live record value and would otherwise mismatch on Date objects.
- */
-function serializeIndexValue(value: unknown): unknown {
-  if (value instanceof Date) return value.toISOString()
-  return value
-}
-
-/**
- * Extract the indexable value for a declaration — a scalar for
- * single-field, or a tuple array for composite. Returns `null` when
- * the value is not indexable (single-field null/undefined, composite
- * with any null/undefined component — the whole composite is skipped
- * if any part is missing).
- */
-function extractIndexValue(
-  record: Record<string, unknown>,
-  def: PersistedIndexDef,
-): unknown {
-  if (def.kind === 'single') {
-    const v = readPersistedValue(record, def.field)
-    return v === undefined || v === null ? null : v
-  }
-  const tuple: unknown[] = []
-  for (const f of def.fields) {
-    const v = readPersistedValue(record, f)
-    if (v === undefined || v === null) return null
-    tuple.push(v)
-  }
-  return tuple
-}
-
-/**
- * Compare the decrypted side-car body's `value` against the live record
- * field value, in the same canonical form used for storage. Handles the
- * Date-is-ISO-string round trip so reconcile doesn't flag a false drift.
- */
-function valuesMatch(stored: unknown, live: unknown): boolean {
-  const serialized = serializeIndexValue(live)
-  if (stored === serialized) return true
-  if (stored === undefined || serialized === undefined) return stored === serialized
-  // JSON-stringify both sides for structural equality on arrays/objects.
-  try {
-    return JSON.stringify(stored) === JSON.stringify(serialized)
-  } catch {
-    return false
+  /**
+   * Bind the {@link IndexingContext} the index-maintenance surface needs. The
+   * eager `cache` Map and the index / unique-constraint / persisted mirrors are
+   * passed by reference (the SAME instances the query path reads, never
+   * copied); the `persistedIndexesLoaded` flag and `ensure*` hydration stay
+   * collection-resident, reached via callbacks.
+   */
+  private indexingContext(): IndexingContext<T> {
+    return {
+      name: this.name,
+      vault: this.vault,
+      adapter: this.adapter,
+      codec: this.codec,
+      cache: this.cache,
+      lazy: this.lazy,
+      emitter: this.emitter,
+      indexes: this.indexes,
+      uniqueConstraints: this.uniqueConstraints,
+      persistedIndexes: this.persistedIndexes,
+      ensureHydrated: () => this.ensureHydrated(),
+      ensurePersistedIndexesLoaded: () => this.ensurePersistedIndexesLoaded(),
+      setPersistedIndexesLoaded: (value) => { this.persistedIndexesLoaded = value },
+    }
   }
 }
