+++ scripts/check-architecture.mjs
@@ -520,7 +520,8 @@ const KERNEL_SURFACE_BUDGET = {
   // Lowered 5433→5401 (Phase 5 A1: deterministic-index extraction): `findByDet`/`queryByDet` moved to `record-keys/deterministic.ts` behind a `DeterministicContext`; Collection keeps thin delegators + a `detContext()` binder.
   // Lowered 5401→5185 (Phase 5 A13: tiers extraction): `putAtTier`/`getAtTier`/`listAtTier`/`elevate`/`demote`/`assertTiersEnabled`/`assertDeclaredTier`/`isElevatorOrOwner`/`_classifySealedShred` moved to `with-audit/tiers/index.ts` behind a `TiersContext`; Collection keeps thin delegators + a `tiersContext()` binder. `cekCache` passed by reference; `emitCrossTierEvent` stays collection-resident (captures `onCrossTierAccess`), reached via callback.
   // Lowered 5185→4962 (Phase 5 A14: search/retrieval extraction): `search`/`flushIndex`/`warmIndex`/`retrieve`/`similarTo` + the build/resolve/retrieveLexical/Semantic/Hybrid helpers moved to `with-lookup/search/collection-facade.ts` behind a `SearchContext`; Collection keeps thin public delegators + a `searchContext()` binder. The eager `cache` Map is passed by reference; `buildPersistedIndexCallbacks` takes a context THUNK (ctor-invoked before `codec` exists, resolved lazily per callback).
-  'packages/hub/src/collection.ts': 4962,
+  // Lowered 4962→4686 (Phase 5 A15: index-maintenance extraction): `rebuildEagerIndexesFromCache`/`rebuildUniqueConstraintsFromCache`/`rebuildIndexes`/`reconcileIndex`/`maintainPersistedIndexesOnPut`/`OnDelete`/`_purgePersistedIndexes` + the readPersistedValue/serializeIndexValue/extractIndexValue/valuesMatch helpers moved to `with-lookup/indexing/collection-facade.ts` behind an `IndexingContext`; Collection keeps thin delegators + an `indexingContext()` binder. The eager `cache` Map + the index/unique/persisted mirrors are passed by reference; `persistedIndexesLoaded` flag + `ensure*` hydration stay collection-resident, reached via callbacks. putInternal/the write path is untouched (still calls the same delegator names).
+  'packages/hub/src/collection.ts': 4686,
   // Bumped 3640→3700 (2026-06-08): deferred-numbering wiring — `sequence()`
   // routing + `runNumberingPass` + the cache-coherent `stamp` closure. The
   // engine itself lives in src/numbering/; only the thin vault call-sites are here.
